Make project's C++ standard more strict
I realized that the compile_commands.json file was using -std=gnu++26 in
the build command, rather than -std=c++26. I looked into it and found
that this could allow for gnu extensions. Turned off extensions in
CMakeLists.txt.

I also found that an older standard can be used as a fallback if the set
standard is not available. Made the specified C++ standard required in
CMake, which will now fail configuration if the requested C++ standard
isn't available.